server_pattern/server/requirement.txt
fastapi
torch
torchvision
uvicorn
pillow
opencv-python
requests